+++ docs/.vitepress/theme/components/NewPost.vue
@@ -1,0 +1,44 @@
+<script setup lang="ts">
+import { computed } from 'vue'
+import { data as posts } from '../../../shared/posts.data'
+import PostBlock from './PostBlock.vue'
+import { usePostSort } from '../composables/usePostSort'
+import { usePostFilter } from '../composables/usePostFilter'
+
+const props = withDefaults(defineProps<{
+  count?: number,
+  showCategory?: boolean
+}>(), {
+  count: -1,
+  showCategory: false
+})
+
+const filterFn = usePostFilter()
+const sortFn = usePostSort()
+
+const newestPosts = computed(() => {
+  if (props.count === -1) {
+    return [...posts].filter(filterFn).sort(sortFn)
+  }
+  return [...posts].filter(filterFn).sort(sortFn).slice(0, props.count)
+})
+</script>
+
+<template>
+  <div :class="$style['new-post-list']">
+    <PostBlock
+      v-for="post in newestPosts"
+      :key="post.url"
+      :post="post"
+      :show-category="props.showCategory"
+    />
+  </div>
+</template>
+
+<style module>
+.new-post-list {
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+</style>